Feat: transform 함수 추가
